+++ CHANGELOG.md
@@ -8,7 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## 0.20.0 - 2023-03-28
+## 0.20.1 - 2023-03-28
+
+### Added
+* Debug progress status
+* Iteration and resume
+
+
+## 0.20.0 - 2023-03-27
 
 ### Added
 * Debug store concept